pyproject.toml
[tool.poetry]
name = "examsbot"
version = "0.1.0"
description = ""
authors = ["Slava Kuryaev <92830706+necrosskull@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "1.0.0"
python-telegram-bot = "20.4"
requests = "^2.31.0"
python-logging-loki = "^0.3.1"
aiofiles = "^23.1.0"


[tool.poetry.group.dev.dependencies]
setuptools = "^68.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
